chatbot-app/server/requirements.txt
uvicorn
fastapi
azure-core
azure-search-documents
python-dotenv
openai
aiohttp

